+++ TORCSBot/torcs-git-master/torcs-git-master/src/drivers/bt/Kernel.cuh
@@ -66,7 +66,7 @@ public:
 
 	static	State*  callKernel(tTrackSeg* segArray, int nTrackSegs, State* initialState,
 							double minXVertex, double maxXVertex, double minYVertex, double maxYVertex,
-							double numIterations,
+							int numIterations,
 							int forwardSegments, double neighborDeltaPos, double neighborDeltaSpeed, double actionSimDeltaTime);
 
 };
